backend/requirements.txt
fastapi
uvicorn
google-generativeai
pandas
openpyxl
python-multipart
python-dotenv
pymupdf
holidays
